changed app icon and renamed assets folder
