+++ src/assets/michael_li_profile_pic.jpg
